+++ app/Http/Controllers/PackageController.php
@@ -210,91 +210,5 @@ class PackageController extends Controller
     }
     
 
-    public function apiPackage()
-    {
-        $packages = Package::all();
-        $package = array_map(function($package){
-            return [
-                "id"=> $package['id'],
-                'name'=> $package['name'],
-                'type'=> $package['type'],
-                'category'=> $package['category'],
-                'price'=> $package['price'],
-                'discount'=> $package['discount'],
-                'description'=> $package['description'],
-                'photo'=> $package['photo'],
-                'list_package'=> array_map(function($detailPackage){
-                    $exam = Exam::where('id', $detailPackage['exam_id'])->get();
-                    foreach ($exam as $item){
-                        return [
-                            'id'=> $detailPackage['id'],
-                            'id_course'=> $item->id,
-                            'name'=> $item->name,
-                            'is_review' => $detailPackage['is_review'],
-                            'amount_acces'=> $detailPackage['amount_acces'],
-                        ];
-                    }
-                    $exam = Course::where('id', $detailPackage['course_id'])->get();
-                    foreach ($exam as $item){
-                        return [
-                            'id'=> $detailPackage['id'],
-                            'id_course'=> $item->id,
-                            'name'=> $item->name,
-                            'is_review' => $detailPackage['is_review'],
-                            'amount_acces'=> $detailPackage['amount_acces'],
-                        ];
-                    }
-                }, PackageDetail::where("package_id",$package['id'])->get()->toArray()),
-            ];
-        },$packages->toArray());
-        return response()->json([
-            "status"=> 200,
-            "message"=> "success",
-            "data"=> $package
-        ]);
-    }
-
-    public function apiPackageById($id)
-    {
-        $packages = Package::where("id", $id)->get();
-        $package = array_map(function($package){
-            return [
-                "id"=> $package['id'],
-                'name'=> $package['name'],
-                'type'=> $package['type'],
-                'category'=> $package['category'],
-                'price'=> $package['price'],
-                'discount'=> $package['discount'],
-                'description'=> $package['description'],
-                'photo'=> $package['photo'],
-                'list_package'=> array_map(function($detailPackage){
-                    $exam = Exam::where('id', $detailPackage['exam_id'])->get();
-                    foreach ($exam as $item){
-                        return [
-                            'id'=> $detailPackage['id'],
-                            'id_course'=> $item->id,
-                            'name'=> $item->name,
-                            'is_review' => $detailPackage['is_review'],
-                            'amount_acces'=> $detailPackage['amount_acces'],
-                        ];
-                    }
-                    $exam = Course::where('id', $detailPackage['course_id'])->get();
-                    foreach ($exam as $item){
-                        return [
-                            'id'=> $detailPackage['id'],
-                            'id_course'=> $item->id,
-                            'name'=> $item->name,
-                            'is_review' => $detailPackage['is_review'],
-                            'amount_acces'=> $detailPackage['amount_acces'],
-                        ];
-                    }
-                }, PackageDetail::where("package_id",$package['id'])->get()->toArray()),
-            ];
-        },$packages->toArray());
-        return response()->json([
-            "status"=> 200,
-            "message"=> "success",
-            "data"=> $package
-        ]);
-    }
+    
 }
